+++ webserver/src/main/java/dev/webserver/tax/TaxRepository.java
@@ -5,11 +5,13 @@ import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
+
 public interface TaxRepository extends CrudRepository<Tax, Long> {
 
     @Transactional
     @Modifying
     @Query("UPDATE tax t SET t.name = :name, t.rate = :rate WHERE t.tax_id = :id")
-    void updateTaxByTaxId(long id, String name, double rate);
+    void updateTaxByTaxId(long id, String name, BigDecimal rate);
 
 }
